+++ RELEASENOTES.md
@@ -26,6 +26,10 @@
         `IllegalArgumentException`
         ([#677](https://github.com/androidx/media/issues/677)).
 *   Video:
+    *   Add workaround for a device issue on Galaxy Tab S7 FE, Chromecast with
+        Google TV, and Lenovo M10 FHD Plus that causes 60fps AVC streams to be
+        marked as unsupported
+        ([#693](https://github.com/androidx/media/issues/693)).
 *   Text:
     *   Remove `ExoplayerCuesDecoder`. Text tracks with `sampleMimeType =
         application/x-media3-cues` are now directly handled by `TextRenderer`